Test: MemberLoation unit test
   - 단위테스트 진행->Member 설정까지 확인
   - service 및 repository 작성하고 테스트 진행예정

+++ src/main/java/com/trablock/demo/domain/location/Coords.java
@@ -11,8 +11,8 @@ import javax.persistence.Embeddable;
 public class Coords {
 
     //private Location location;
-    private Float latitude;
-    private Float longitude;
+    private double latitude;
+    private double longitude;
 
     protected Coords() {
     }
